gh: irc: restrict Tests notifs
Due to a bug somewhere, we get notifications for builds that have not
been started, e.g.

  https://github.com/multipath-tcp/mptcp_net-next/runs/3409839364?check_suite_focus=true

It looks like the 'conclusion' field for those is always set to
'neutral' while it should be either success or failure (or timed_out,
etc.).

We can then filter out 'neutral' ones to avoid useless notifications
like:

  New Tests job validating for-review (by matttbe) just ended:
  - {"code":5,"message":"NOT_FOUND: Task '6133726292738048' not found!"}: https://cirrus-ci.com/build/6133726292738048
  - {"code":404,"message":"HTTP 404 Not Found"}:

An alternative would be to look at the 'head branch' but just in case
we want to validate more branches/tags later (e.g. from patchew), best
not to go to this direction if possible.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>