cirrus: limit cpu/mem per test
Because we have a limitation per repo (16 CPUs), best to reduce to 4 to
allow 2x2 tests in parallel.

According to graphs on Cirrus reports and thanks to the cache, we don't
really use more than 2 or 3 cores at the same time most of the time.
When we use more, that's often because of the compilation but ccache is
helping a lot here.

We could limit later to 2 CPUs to allow 4 tests in parallel but I don't
think we often have that many submissions at the same time.

While at it, we can also reduce the memory to 4GB as we don't really use
more than 3GB per test.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>