gh: irc: check-runs can have extra entries
Sometimes, an extra extry is present in check-runs results, e.g.

  https://api.github.com/repos/multipath-tcp/mptcp_net-next/check-suites/4038566338/check-runs

We can find a "Build Parsing Results" entry pointing to the build
address and not the tasks. We need to filter out this and only take
tasks where the conclusion is not "neutral".

While at it, also get the JSON with curl and parse it with jq for
advanced filtering, all in one step.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>