gh: notif: add email notif for patchew
When a patch is sent to the ML, a tag is created on GH which starts
actions: builds and tests. But no notification is sent at the end.

Here we fill this gap by sending emails to the ML in reply to the last
patch of the series.

Note that for the moment, we have to use a custom version of
'action-send-mail' to support 'In Reply To' tag.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>