selftests: mptcp: more stable simult_flows tests
Currently the simult_flows.sh self-tests are not very
stables, expecially when running on slow VM.

The tests mesures runtime for transfers on multiple subflows
and check that the time is nearby the theoretical maximum.

The current test infra introduces a bit of jitter in test
runtime, due to multiple explicit delays. Additionally the
runtime is measured by the shell script wrapper. On slow
VM, the script overhead is measurable and subject to relevant
jitter.

One solution to make the test more stable would be adding more
slack to the expected time; that could possibly hide reall
regressions. Instead move the measurement inside the command
doint the transfer, and drop most unneeded sleep.

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Message-Id: <9d66028a72b6807d4dc3397bb70f028cbc78161d.1634310418.git.pabeni@redhat.com>