cirrus: increase timeout to 90m instead of 60
Because we recently reduce the number of allocated CPU, it can take more
time to compile the kernel and other preparation tasks. We are now
reaching this global timeout when modifications in the upstream kernel
deprecate an important part of ccache like we just had last night.

When we reach the global timeout, everything stop and we don't even have
artifacts created with the conclusion and other info we use in
notifications.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>