gh: build: try to fix ccache restoration
It looks like the saving is OK but not the restoration.

I suspect the fact it uses the longest string up to the last dash to
create the key. So I suppressed an extra dash after the key we want to
restore and put the 'restore-keys' in a list.

We will see...

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>