gh: notif: temp disable email for tests
Instead, only keep IRC notif for the moment.

The issue is that these tests are still quite unstable. Then their value
is quite low and it is not a useful info to share on the ML for the
moment.

But we keep a notification via IRC (and just to me via email) not to
forget about them.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>