gh: notif: email: fix in-reply-to for reapplied patches
It seems Patchew re-applies patches after someone sent a tag (Acked-by,
etc.) on the ML. But this tag is added at the end, breaking the way we
get the Message-Id.

We should not have Patchew re-applying the patches after a tag is sent
but anyway, safer to grep for Message-Id, that's what we want.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>