gh: build: make key different all the time
With run_id (unique per triggered build) and run_attempt (unique per run
of a triggered build), we are sure to never match the input key.

The reason why it is important to do that is that if we match the input
key -- e.g. re-run -- , the cache system will not publish new cache
while ccache has certainly changed some stuff.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>