gh: build: finally re-use cache
The doc is not very clear about that but in short, if we build tags, we
cannot re-use cache except if we re-build the tag.

In our case, we want to re-use the cache. So as a workaround, we build
the 'export' branch and we retrieve the linked tag later with Github's
API. Not nice but if that's how Github wants us to use their resources.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>