mptcp: send MP_FAIL with MP_RST
As described in RFC8684, 3.7 Fallback:

'''
Therefore, it is not possible to recover the subflow, and the affected
subflow must be immediately closed with a RST that includes an MP_FAIL
option (Figure 16), which defines the data sequence number at the start
of the segment (defined by the Data Sequence Mapping) that had the
checksum failure.
'''

MP_FAIL could be sent with MP_RST at the same time.

This patch fixed it.

Fixes: c3b546f7292ac ("mptcp: implement fastclose xmit path")
Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <2ba9856233bd45ffe99f01566c2772b0dd12a501.1637630218.git.geliang.tang@suse.com>