selftests: mptcp: implement pm_nl_set_limits
This patch implemented a new function named pm_nl_set_limits(), wraped
the PM netlink commands 'ip mptcp' and 'pm_nl_ctl' in it, and used a new
argument 'ip_mptcp' to choose which one to use for the PM limits
setting.

Used this wrapper in all the selftests in mptcp_join.sh instead of using
the pm_nl_ctl commands directly.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <f4d5d1b1122215ce7a1bb723b96206852e83e01f.1641534346.git.geliang.tang@suse.com>