selftests: mptcp: implement pm_nl_add_endpoint
This patch implemented a new function named pm_nl_add_endpoint(), wraped
the PM netlink commands 'ip mptcp' and 'pm_nl_ctl' in it, and used a new
argument 'ip_mptcp' to choose which one to use for adding the PM endpoint.

Used this wrapper in all the selftests in mptcp_join.sh instead of using
the pm_nl_ctl commands directly.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <73fc4696a94038df411a0b1c4ed329ffb3abcd57.1641534346.git.geliang.tang@suse.com>