selftests: mptcp: implement pm_nl_show_endpoint
This patch implemented a new function named pm_nl_show_endpoint(), wraped
the PM netlink commands 'ip mptcp' and 'pm_nl_ctl' in it, and used a new
argument 'ip_mptcp' to choose which one to use to show all the PM endpoints.

Used this wrapper in all the selftests in mptcp_join.sh instead of using
the pm_nl_ctl commands directly.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <2663a2bb82272c43ab1578d5f272f937bd225225.1641534346.git.geliang.tang@suse.com>