selftests: mptcp: implement pm_nl_del_endpoint
This patch implemented a new function named pm_nl_del_endpoint(), wraped
the PM netlink commands 'ip mptcp' and 'pm_nl_ctl' in it, and used a new
argument 'ip_mptcp' to choose which one to use for deleting the PM
endpoint.

Used this wrapper in all the selftests in mptcp_join.sh instead of using
the pm_nl_ctl commands directly.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <74ccf527130bfbebe3e92ab5b6328e502c880f91.1641535880.git.geliang.tang@suse.com>