selftests: mptcp: implement pm_nl_flush_endpoint
This patch implemented a new function named pm_nl_flush_endpoint(), wraped
the PM netlink commands 'ip mptcp' and 'pm_nl_ctl' in it, and used a new
argument 'ip_mptcp' to choose which one to use for flushing the PM
endpoints.

Used this wrapper in all the selftests in mptcp_join.sh instead of using
the pm_nl_ctl commands directly.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <62c2e9c36e5de33653ffbaf3a282c24f46d6f873.1641535880.git.geliang.tang@suse.com>