selftests: mptcp: implement pm_nl_set_endpoint
This patch implemented a new function named pm_nl_set_endpoint(), wraped
the PM netlink commands 'ip mptcp' and 'pm_nl_ctl' in it, and used a new
argument 'ip_mptcp' to choose which one to use to set the flags of the PM
endpoint.

Used this wrapper in all the selftests in mptcp_join.sh instead of using
the pm_nl_ctl commands directly.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <f7b86a8bb99769cefd4e9cd7673b0228479a498a.1641535880.git.geliang.tang@suse.com>