selftests: mptcp: new approach of getting ID
This patch implemented a more flexible approach to get the address ID
from the PM dump output.

The original 'pos+=5' in the remoing tests only works for the simple
case of 'pm_nl_ctl show':

id 1 flags subflow 10.0.1.1

It dosen't work for the outputs with dev or port:

id 1 flags signal 10.0.2.1 10100
id 1 flags subflow dev ns2eth3 10.0.3.2

This patch rewrites it to fit for more cases and the output of
'ip mptcp'.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <14e547f9691ee85c0e34a0c8c87a3e6f19612b22.1642154579.git.geliang.tang@suse.com>