Squash to "selftests: mptcp: add wrapper for setting flags"
Please update the commit message:

'''
This patch implemented a new function named pm_nl_set_endpoint(), wrap
the PM netlink commands 'ip mptcp endpoint change flags' and 'pm_nl_ctl
set flags' in it, use a new argument 'ip_mptcp' to choose which one to
use to set the flags of the PM endpoint. Randomly choose the 'addr' or
'id' as arguments for each command.

Parse the address ID from the PM dump output as well as the address and
port number, pass them to this wrapper in do_transfer() instead of using
the pm_nl_ctl command directly.
'''

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <f7696ad95afc97a4c78443ff700c588b600cc822.1642749049.git.geliang.tang@suse.com>