selftests: mptcp: increase timeout to 1 hour
With the increase number of tests, on a slow public CI with kernel debug
settings, it has been observed it can easily and regularly take up to 45
minutes to execute all tests in this very slow environment. The slowest
test, mptcp_join.sh takes ~30 minutes in this case.

In such environments, the selftests timeout set in the 'settings' file
is disabled. But because other similar environments can be used to run
MPTCP selftests, it is certainly better to increase the timeout for
everybody else to 1h.

Most of the time, running all MPTCP selftests takes around 5 to 10
minutes on decent environments and the slowest test takes less than 10
minutes, the current timeout.

The timeout could be disabled but it is always good to have an extra
safeguard, just in case.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20220207100022.3936015-1-matthieu.baerts@tessares.net>