mptcp: add echo bit in mptcp_addr_info
The echo bit is one of the important properties of a MPTCP address. It
makes sense to put it into struct mptcp_addr_info with others properties
like the address family, the id number and the port number too.

This patch added the echo bit in struct mptcp_addr_info. Use this instead
of using the struct member echo in struct mptcp_options_received.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <e3236d76eae1f61afd43bcb1ad08f79f859bddb4.1644318561.git.geliang.tang@suse.com>