selftests: mptcp: add fastclose testcases
This patch added the self tests for MP_FASTCLOSE. Added a new argment
fastclose for the function do_transfer(). Passed the extra argments '-I 2'
and '-I 3' to the mptcp_connect commands to disconnect the connections to
trigger the MP_FASTCLOSE sending and receiving. Used chk_fclose_nr to check
the MP_FASTCLOSE mibs and used chk_rst_nr to check the MP_RST mibs.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <9eaa5384986ccd5d42efcceab5fe722cbb876382.1644682054.git.geliang.tang@suse.com>