Squash-to: "mptcp: more careful RM_ADDR generation"
When the self-tests flushed the subflows on both sides of
the MPTCP connections, the number of subflows actually deleted
on each side is unreliable.

The test code always flushes  first the client side, but the
actual delete operations is performed asynchronously, and one
on more subflows could be processed after that the other end
already deleted it.

The grand total of the deleted subflows is expected to be
in the [nr_subflows: 2* nr_subflows] range.

Address the issue checking the total number of deleted subflows
in the simult flush cases and enforcing the above constraint

Signed-off-by: Paolo Abeni <pabeni@redhat.com>
Tested-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>