Squash to "selftests: mptcp: join: allow running -cCi"
The recommended way with getopt is first parse all options and then act.

This allows to do some actions in priority, e.g. display the help menu
and stop.

But also some global variables changing the behaviour of this selftests
-- like the ones behind -cCi options -- can be set before running the
different tests. By doing that, we can also avoid long and unreadable
regex.

This can be squashed in "selftests: mptcp: join: allow running -cCi" and
the commit message can be appended with what is above.

Reviewed-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20220224135619.2523272-3-matthieu.baerts@tessares.net>