tcp: Allow tcp_check_req() to return more status values
tcp_check_req() can already use an output parameter to set req_status,
indicating whether the current thread has lost a race to process the req.

Convert this bool pointer to instead use an enum, so tcp_check_req() can
look for and set status values indicating other situations. This is used
in a subsequent commit to check for a MPTCP MP_JOIN header. There is no
functional change intended in this patch.

Signed-off-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Message-Id: <20220224232226.259304-3-mathew.j.martineau@linux.intel.com>