cirrus: increase timeout from 90 to 115min
With the recent modifications to use 'config.debug', it takes around 60
min to run the tests. But if there is a lot to recompile, the clone and
compilation on a busy VM can probably take ~30 minutes.

To be on the safe side, we increase the timeout by 25 minutes, 5 under
the hard limit.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>