selftests: mptcp: check MP_FAIL response mibs
This patch extended chk_fail_nr to check the MP_FAIL response mibs.

Added a new argument invert for chk_fail_nr to allow it can check the
MP_FAIL Tx and Rx mibs from the opposite direction.

When the infinite map was received before the MP_FAIL response, the
response will be lost. A '-' can be added into fail_tx or fail_rx to
represent that MP_FAIL response Tx or Rx can be lost when doing the
checks.

Signed-off-by: Geliang Tang <geliang.tang@suse.com>
Message-Id: <f278d5dda12571f464f4b97da68a6627de3eea95.1646829006.git.geliang.tang@suse.com>