selftests: mptcp: add the MP_FAIL testcases
Add the test cases for MP_FAIL, the multiple subflows test for the
MP_RST case and the single subflow one for the infinite mapping case.
The former uses the test_linkfail value to make 1024KB test files,
and the latter 128KB.

Add a new function reset_with_fail(), in it use 'iptables' and 'tc
action pedit' rules to produce the bit flips to trigger the checksum
failures. Add a new function pedit_action_pkts() to get the numbers
of the packets edited by the tc pedit actions. Set validate_checksum
to enable checksums for the MP_FAIL tests without passing the '-C'
argument.

Also add the needed kernel configures in the selftests config file.

Suggested-by: Davide Caratti <dcaratti@redhat.com>
Co-developed-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Reviewed-by: Mat Martineau <mathew.j.martineau@linux.intel.com>
Signed-off-by: Geliang Tang <geliang.tang@suse.com>