Squash to "mptcp: handle local addrs announced by userspace PMs"
Move mptcp_pm_addr_entry structure to protocol.h as this commit exports
a new function having a parameter using this structure.

This was causing the CI to complain about this commit (and the 2 next
ones). Fixing it here instead of modifying the CI not to complain about
that.

Note: compared to what Kishen did in commit "mptcp: netlink: Add
MPTCP_PM_CMD_ANNOUNCE", I moved the new structure with other mptcp_pm
ones.

Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>