Squash to "selftests: mptcp: add subflow limits test-cases"
The test -- "delete and re-add" introduced by the above commit -- is
failing on the different CIs with a debug kernel config. It is also
failing without the debug config but it is less frequent, only on the
public CI.

Increasing the packet size from 1KB to 4KB makes the kernel sending
multiple packets and it seems to fix the issues reported by the
different CIs.

Note: Paolo shared the diff on IRC without a commit message. I prefer to
apply it like that (without his SoB) so we can stop ignoring CI's
output.

Closes: https://github.com/multipath-tcp/mptcp_net-next/issues/274
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Message-Id: <20220506171819.1104074-1-matthieu.baerts@tessares.net>